Ai-Integration/requirements.txt
numpy==1.21.5
pandas==1.3.5
scikit-learn==1.0.2
tensorflow==2.8.0
keras==2.8.0
joblib==1.1.0
nltk==3.6.7
scipy==1.7.3
matplotlib==3.5.1
seaborn==0.11.2
python-dateutil==2.8.2
Flask==2.1.0
torch==1.9.0
transformers==4.11.3
langdetect==1.0.9
redis==3.5.3
sentry-sdk[flask]==1.5.0
Flask-Limiter==2.4.0
schedule==1.1.0
gunicorn==20.1.0
requests==2.26.0
Flask==2.1.0
torch==1.9.0
transformers==4.11.3
langdetect==1.0.9
redis==3.5.3
sentry-sdk[flask]==1.5.0
Flask-Limiter==2.4.0
schedule==1.1.0
gunicorn==20.1.0
requests==2.26.0
PyPDF2==2.10.5
python-docx==0.8.11
SpeechRecognition==3.8.1
pydub==0.25.1
gTTS==2.2.3
moviepy==1.0.3
Werkzeug==2.0.2
workalendar==14.1.0
sqlalchemy==1.4.31
Flask==2.1.0
Werkzeug==2.0.3
torch==1.9.0
transformers==4.11.3
nltk==3.6.3
scikit-learn==0.24.2
numpy==1.21.2
opencv-python==4.5.3.56
pytesseract==0.3.8
pdf2image==1.16.0
Pillow==8.3.2
SpeechRecognition==3.8.1
pydub==0.25.1
flask-limiter==1.4
Flask-Caching==1.10.1
prometheus-flask-exporter==0.18.0
gunicorn==20.1.0
pytest==6.2.5
logging==0.4.9.6
Flask-Talisman==0.8.1
flasgger==0.9.5
black==21.9b0
flake8==3.9.2
aiohttp==3.8.1
asyncio==3.4.3
openai==0.27.0
PyMuPDF==1.19.6
profanity-check==1.0.3
SQLAlchemy==1.4.31
rq==1.10.1
rq-scheduler==0.11.0
boto3==1.20.26
cachetools==4.2.4
tenacity==8.0.1
Jinja2==3.0.3
reportlab==3.6.5
spacy==3.2.1
aiofiles==0.8.0
python-dotenv==0.19.2







